+++ R/SuperpixelImageSegmentation.R
@@ -1,0 +1,3 @@
+#' @useDynLib SuperpixelImageSegmentation, .registration = TRUE
+#' @importFrom Rcpp evalCpp
+NULL